requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
PyPDF2>=3.0.0

# Embeddings
sentence-transformers>=2.3.0

# Vector Database
chromadb>=0.4.0

# Data processing
pandas>=2.1.0

# API (Phase 9)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# LLM (Phase 6)
groq>=0.5.0

# Utilities
python-dotenv>=1.0.0
apscheduler>=3.10.0
pytz>=2023.3
